Convert stream takes input stream and converts it to string
Changed root of routes to be just a forward slash, that way it can take
in any port at start and boot the server.

Server currently hard codes which resource to return which is
index.html  and turns it into a string in the server class to prepare
for the response.

Added an html file for index.html