Post request is handle and body is updated
action for form is called when a /form url is requested, if a get
request is called the response includes a 200 OK and the response body
is simply a blank html page except an h1 stating the title.

A post request finds the file, updates that file, then sends only the
response header back.

When a get request is called again the user sees the updated page with
that data.