server calls handler method called handle request and response
Refactor to try to have similar levels of abstraction for methods.
Handle request hast to parse all the information through input streams,
response uses this parsed information to then retrieve a string that is
the response through finding the route, going to a controller action,
and returning a string to send back to the server.